+++ PulseAudioSource.hpp
@@ -6,7 +6,7 @@ class PulseAudioSource : public AudioSource {
   public:
     PulseAudioSource();
     ~PulseAudioSource();
-    virtual void read(float *samples, size_t num);
+    virtual void read(double *samples, size_t num);
     virtual unsigned int getSampleRate();
   private:
     pa_simple *s;
